"던전 플레이 로직 완성"
Fix : "던전 플레이 로직 완성"

Body : 던전플레이시 골드가 포함되지 않는 부분, 체력이 감소되지 않는 부분을 프로퍼티 charcter를 통해 해결

Footer :

+++ TRPG/Characters.cs
@@ -10,14 +10,15 @@ namespace TRPG
     {
         public enum jobType { 전사, 궁수, 마법사 }
         public string Name { get; set; }
-        public float Health { get; set; }
-        public float Attack { get; set; }
+        public int Health { get; set; }
+        public int Attack { get; set; }
         public float Level { get; set; }
-        public float Armor { get; set; }
-        public float Gold { get; set; }
+        public int Armor { get; set; }
+        public int Gold { get; set; }
         bool IsDead => Health <= 0;
         public jobType Job { get; set; }
-        public float maxHealth { get; set; }
+        public int maxHealth { get; set; }
+        
         
         public Characters(string name, jobType job)
         {
@@ -28,6 +29,7 @@ namespace TRPG
             Attack = 20;  //아이템 장착까지 고려한 공격력을 0으로 초기화
             Gold = 1500;
             Armor = 5;   //아이템 장착까지 고려한 방어력을 0으로 초기화
+            
         }
         public static Characters CreateCharacters(string name, jobType job) //enum으로 받은 직업을 통해서
                                                                //캐릭터 생성시 매개변수로 이름과 직업을 선택하도록
